change the ‘do not update the user’ first run error threshold
if a run has errored requests for a tenth of more of the requests sent, the client will not be updated with the subreddit list.

this is to prevent a sudden incessant spamming of ‘sub gone private’ notifications when reddit starts allowing the requests again, as the script will have assumed the subs are public.

yeah in an ideal world the subreddit status list would probably be on a db, but i don’t think i’ve got enough time before the end of the blackout to set one up and get it working. maybe if it continues or there’s another blackout, who knows. (it would also prevent my cpu from being damn near overwhelmed when it refreshes the list of 7000+ participating subs.

this limit was previously set to 1/2 but was changed after an **extremely close call** where reddit stopped rate limiting suddenly in the middle of an `updateStatus` call, and it was literally by a hair’s breadth that around 1000+ clients didn’t get a sudden spam of 3500 false ‘gone private’ notifs.