remove clientside filtering list
as part of these changes, we can move the filter list onto the back-end. (even with auto-filtering, i think there’ll still be a use for a filter list, for those that are continuously spamming - to avoid even seeing X alerts for them every hour). that way, it keeps things simple, with the flag coming from the server being the single arbiter on whether or not an alert is displayed. (it also avoids necessitating a forced refresh on all clients when the filter list is updated.)